# api/swagger.yaml
swagger: "2.0"
info:
  description: "Esta es la referencia de la REST API, se hace documentación de los endpoints de nuestra aplicacion"
  version: "1.0"
  title: "Candidatos API"
  termsOfService: "https://opensource.org/licenses/MIT"
  contact:
    email: "ernestopm20@gmail.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:8005"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /api/v1/python/predictTweet:
    post:
      tags:
      - "python"
      summary: "Regresa el centroide al que pertenece el tweet pasado en el body"
      description: "Se ejecuta el script en python para predecir ese tweet" 
      operationId: "predictTweet"
      consumes:
        - "application/json"
      produces:
        - "application/json"
  
      parameters:
        - name: body 
          in: body
          description: "Contiene el tweet a predecir por ejemplo: Hola como estas y el frontendID de cada candidato"
          required: true
          schema:
            $ref: '#/definitions/pythonBody'
            
      responses:
        400:
          description: "Datos faltantes, probablemente falta la query en el body"
        200:
          description: "Operación Exitosa"
          schema:
            title: "Cosa"
            type: "object"
            properties:
              results:
                type: "array"
                items: {
                  type: "string",
                }
             
        500:
          description: "Error en el servidor"          
definitions:
  pythonBody:
    type: "object"
    properties:
      tweet:
        type: "string"
      candidateFrontendId:
        type: "string"
...